# Use upstream XLA concurrency utilities