--- a/torch_xla/csrc/runtime/BUILD
+++ b/torch_xla/csrc/runtime/BUILD
@@ -286,6 +286,7 @@ cc_library(
     deps = [
         ":metrics",
         ":tf_logging",
+        "@tsl//tsl/platform:env"
     ],
 )