--- a/torch_xla/csrc/runtime/thread_pool.cc
+++ b/torch_xla/csrc/runtime/thread_pool.cc
@@ -16,7 +16,8 @@ namespace runtime {
 void Schedule(std::function<void()> fn) {
   static size_t num_threads = sys_util::GetEnvInt(
       "XLA_THREAD_POOL_SIZE", std::thread::hardware_concurrency());
-  static tsl::thread::ThreadPool pool(tsl::Env::Default(), "pytorchxla", num_threads);
+  static tsl::thread::ThreadPool pool(tsl::Env::Default(), "pytorchxla",
+                                      num_threads);
   pool.Schedule(std::move(fn));
 }